+++ internal/aws/imagebuilder/image_recipe_resource_gen.go
@@ -92,6 +92,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			// CloudFormation resource type schema:
 			// {
 			//   "description": "The block device mappings to apply when creating images from this recipe.",
+			//   "insertionOrder": false,
 			//   "items": {
 			//     "additionalProperties": false,
 			//     "description": "Defines block device mappings for the instance used to configure your image. ",
@@ -125,7 +126,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			//             "type": "string"
 			//           },
 			//           "Throughput": {
-			//             "description": "For GP3 volumes only ? The throughput in MiB/s that the volume supports.",
+			//             "description": "For GP3 volumes only - The throughput in MiB/s that the volume supports.",
 			//             "type": "integer"
 			//           },
 			//           "VolumeSize": {
@@ -207,7 +208,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 								},
 								"throughput": {
 									// Property: Throughput
-									Description: "For GP3 volumes only ? The throughput in MiB/s that the volume supports.",
+									Description: "For GP3 volumes only - The throughput in MiB/s that the volume supports.",
 									Type:        types.Int64Type,
 									Optional:    true,
 								},
@@ -256,6 +257,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			Optional: true,
 			Computed: true,
 			PlanModifiers: []tfsdk.AttributePlanModifier{
+				Multiset(),
 				tfsdk.UseStateForUnknown(),
 				tfsdk.RequiresReplace(),
 			},
@@ -265,6 +267,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			// CloudFormation resource type schema:
 			// {
 			//   "description": "The components of the image recipe.",
+			//   "insertionOrder": true,
 			//   "items": {
 			//     "additionalProperties": false,
 			//     "description": "Configuration details of the component.",
@@ -275,6 +278,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			//       },
 			//       "Parameters": {
 			//         "description": "A group of parameter settings that are used to configure the component for a specific recipe.",
+			//         "insertionOrder": false,
 			//         "items": {
 			//           "additionalProperties": false,
 			//           "description": "Contains a key/value pair that sets the named component parameter.",
@@ -285,6 +289,7 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			//             },
 			//             "Value": {
 			//               "description": "Sets the value for the named component parameter.",
+			//               "insertionOrder": true,
 			//               "items": {
 			//                 "type": "string"
 			//               },
@@ -334,6 +339,9 @@ func imageRecipeResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 							tfsdk.ListNestedAttributesOptions{},
 						),
 						Optional: true,
+						PlanModifiers: []tfsdk.AttributePlanModifier{
+							Multiset(),
+						},
 					},
 				},
 				tfsdk.ListNestedAttributesOptions{},